Add MapHelp command
MapHelp makes it easier to lookup what a command does. Most of the time,
it returns whatever you've mapped the input key sequence to, but if it's
built-in it shows vimdoc for the map (instead of an error like :map).

Useful whenever you know you use something a lot, but can't remember
where it comes from.

Also includes a bang variant to use :Verbose instead of verbose.

Passing no modes will search all modes unlike :map or maparg(). This is
the most useful use case. You can pass a mode to pinpoint entries in
vimdoc or for Verbose.